Use character-by-character comparison for comparison operators
This follows the proposal in section [uri.header-synopsis.equality-comparison].

Otherwise the type would do expensive allocation due to normalization of the
URIs when used in containers or algorithms.

The operator functions have been made friend functions of the uri class
to take advantage of comparing the underlying string_view instead of
re-implementing the behavior based on iterators.